+++ frontend/src/api.ts
@@ -7,3 +7,4 @@ export const NAVER_AUTH_URI = `https://nid.naver.com/oauth2.0/authorize?response
 }&state=${`${Math.random().toString(36).substr(3, 14)}`}&redirect_uri=${NAVER_REDIRECT_URI}`;
 
 export const NATIONAL_TAX_SERVICE_BASE_URL = `https://api.odcloud.kr/api/nts-businessman/v1`;
+export const NATIONAL_TAX_SERVICE_STATUS_API_URL = `${NATIONAL_TAX_SERVICE_BASE_URL}/status?serviceKey=${process.env.REACT_APP_NATIONAL_TAX_SERVICE_KEY}`;
